Let player throw grabbable stuff
Adds a GrabbableBehavior script that can be attached to boxes, bags etc. The ground layer must be set and the object must be tagged as Grabbable like before. This script has functions to pick up the grabbable, disabling its physics, and to throw it, re-enabling its physics and passing in an initial velocity. Also adds logic to the player controller to make use of these functions.