Made pressure switches 20 times cooler
Refactored door and switch code.

Doors now inherit from an OnOffSwitchReceiver class which has functions for OnSwitchDown and OnSwitchUp. Pressure switches target this class, so other things aside from doors can use it to be controlled by the same switches, and other things besides switches could implement it to control doors.

The player, box and bag have a new script called SwitchWeight. This script allows a weight value to be defined and recursively calculated by checking the GameObject's children, to report the total weight for cases such as a player with a weight carrying a bag with a weight.

Replaced DoorSwitchBehavior.cs with a new PressureSwitchBehavior.cs. These switches have three states: unpressed, partially pressed and fully pressed. When the switch becomes fully pressed, it sends an OnSwitchDown call to its OnOffSwitchReceiver. When it becomes not fully pressed, it sends OnSwitchUp. In order to be fully pressed, the switch must have a cumulative weight from objects with SwitchWeight scripts greater than the full press threshold defined in the inspector. Half presses are purely aesthetic and can also be defined with a threshold, to give the player feedback that they must find something heavier. Pressure Switch Behavior also includes an option to disable switches and lock them into a pressed state once they have been initially pressed, and a logging switch that will make them print the number and cumulative weight of objects on them every time they receive a collision update.

Also made the switch into a prefab and defined a new HasGravity layer for the player, boxes and bags.