Grabby Cops
Cops now walk back and forth. There is an invisible Cop Blocker prefab that can be placed to make cops turn around without interfering with the player. Cops don't hurt the player or the bag, but will grab the bag if they touch it while it's stationary and unattended. If they do, they must be taken out with a flying projectile to get it back. Dead cops can be grabbed and used as weights/projectiles themselves. TODO: make cops hurt the player (currently disabled, going to do it differently), make cops speed up when they see something interesting ahead.