Make voiceovers only play once even after death
Adds a SingleRunBehavior that VoiceOverTriggers inherit from, and a SingleRunManager that SingleRunBehaviors communicate with. The manager is persistent via DontDestroyOnLoad and stores the data for each behavior so that they can safely be reset when the scene is loading, but the scene can still hold as many as it needs. There are probably easier ways to do this.